+++ lib/src/cull_trim_path.dart
@@ -1,9 +1,0 @@
-// import 'dart:ui';
-
-// Path trimPath(Path source, double length, {bool firstOnly = true}) {}
-
-// Path trimPathStart(Path source, double length, {bool firstOnly = true}) {}
-
-// Path trimPathEnd(Path source, double length) {}
-
-// Path cullPath(Path source, Rect bounds) {}